Minor cleanup
Add lombok!
