+++ build.gradle
@@ -17,6 +17,11 @@ dependencies {
     compile group: 'org.json', name: 'json', version: '20190722'
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
+
+    compileOnly 'org.projectlombok:lombok:1.16.16'
+    annotationProcessor 'org.projectlombok:lombok:1.18.6'
+
+    implementation 'com.eclipsesource.j2v8:j2v8_macosx_x86_64:4.6.0'
 }
 
 // See https://github.com/JetBrains/gradle-intellij-plugin/
@@ -25,6 +30,7 @@ intellij {
     version '2019.3.4'
     type 'IU'
 }
+
 patchPluginXml {
     changeNotes """
       Add change notes here.<br>
