+++ package.json
@@ -1,0 +1,8 @@
+{
+	"name": "codeshovel",
+	"version": "",
+	"license": "UNLICENSED",
+	"devDependencies": {
+		"typescript": "^3.9.2"
+	}
+}
