+++ src/main/java/com/lucasaz/intellij/AssertionGeneration/dto/Selected.java
@@ -1,0 +1,16 @@
+package com.lucasaz.intellij.AssertionGeneration.dto;
+
+import lombok.*;
+
+@Getter
+@AllArgsConstructor
+@Builder
+@ToString
+public class Selected {
+    int line;
+    int col;
+    String selected;
+    String tsFilePath;
+    String originalFile;
+    String whitespace;
+}
