+++ src/main/java/com/lucasaz/intellij/AssertionGeneration/indices/AssertionGenerationSettingsConfigurable.java
@@ -1,9 +1,11 @@
-package com.lucasaz.intellij.AssertionGeneration;
+package com.lucasaz.intellij.AssertionGeneration.indices;
 
 import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.options.SearchableConfigurable;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectUtil;
+import com.lucasaz.intellij.AssertionGeneration.AssertionGenerationSettingsForm;
+import com.lucasaz.intellij.AssertionGeneration.util.Util;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
