+++ src/main/java/com/lucasaz/intellij/AssertionGeneration/visitors/IVisitor.java
@@ -1,0 +1,6 @@
+package com.lucasaz.intellij.AssertionGeneration.visitors;
+
+public interface IVisitor<T> {
+	void visit(T visitable);
+	void close();
+}
